+++ silkwood/src/render/shaders/outline_shader.wgsl
