+++ silkwood/src/render/shaders/primitive_shader.wgsl
