+++ silkwood/src/render/shaders/texture_shader.wgsl
