+++ silkwood/src/render/ui/mod.rs
@@ -1,0 +1,3 @@
+//! Structs for building the ui for parts of the game
+
+pub mod text;
