+++ src/lib.rs
@@ -1,5 +1,4 @@
 pub mod app;
 pub mod parser;
-pub mod render;
 pub mod settings;
 pub mod track;
